Enable Mobile Browser SSO for Native mobile Applications.
The workday mobile application works in one of twos way when signing in to an identity provider. It can either initiate a browser based sign on for which it will transition from the native application to the default browser. This breaks conditions access compliance because this method does not pass all device details. 
The other method is to sign in via the native application. This allows the application to pass the device details to Azure and Azure is able to determine that the device is managed, compliant etc.