{URGENT} REQUIRE_USER_MATCH values in Upper Case
The document has the REQUIRE_USER_MATCH  RegKey value as "True" or "False". The Reg Key is case sensitive and when incorrectly configured as "True" will allow an already registered user to remove all authentication methods and bypass MFA. A user who keeps their registered authentication methods continue to get prompted for MFA.
It is imperative that the article is updated to reflect the case sensitivity.